+++ Makefile
@@ -133,7 +133,7 @@ protobuild:
 	@protoc-go-inject-tag -input=./internal/kms/store/oidc_key.pb.go		
 	@protoc-go-inject-tag -input=./internal/target/store/target.pb.go
 	@protoc-go-inject-tag -input=./internal/auth/oidc/store/oidc.pb.go
-	@protoc-go-inject-tag -input=./internal/job/store/job.pb.go
+	@protoc-go-inject-tag -input=./internal/scheduler/job/store/job.pb.go
 
 	@rm -R ${TMP_DIR}
