+++ internal/scheduler/job/additional_verification_test.go
